+++ modules/core/src/core.module.ts
@@ -5,12 +5,13 @@ import { DynamicFormService } from "./service/dynamic-form.service";
 import { DynamicFormAutoFillService } from "./service/dynamic-form-autofill.service";
 import { DynamicFormRelationService } from "./service/dynamic-form-relation.service";
 import { DynamicIdDirective } from "./directive/dynamic-id.directive";
+import { DynamicTemplateDirective } from "./directive/dynamic-template.directive";
 
 @NgModule({
 
     imports: [CommonModule, ReactiveFormsModule],
-    declarations: [DynamicIdDirective],
-    exports: [DynamicIdDirective]
+    declarations: [DynamicIdDirective, DynamicTemplateDirective],
+    exports: [DynamicIdDirective, DynamicTemplateDirective]
 })
 
 export class DynamicFormsCoreModule {