+++ modules/core/src/directive/dynamic-template.directive.ts
@@ -1,0 +1,12 @@
+import { Directive, Input, TemplateRef } from "@angular/core";
+
+@Directive({
+    selector: "template[modelId]"
+})
+
+export class DynamicTemplateDirective {
+
+    @Input() modelId: string;
+
+    constructor(public templateRef: TemplateRef<any>) {}
+}