+++ modules/ui-material/src/dynamic-form-material.component.ts
@@ -1,11 +1,12 @@
-import { Component, Input, Output, EventEmitter, ViewChild, ContentChild, TemplateRef } from "@angular/core";
+import { Component, Input, Output, EventEmitter, QueryList, ViewChild, ContentChildren } from "@angular/core";
 import { FormGroup } from "@angular/forms";
 import { MdCheckbox, MdInputContainer, MdRadioGroup, MdSelect, MdSlider, MdSlideToggle } from "@angular/material";
 import {
     DynamicFormControlComponent,
     DynamicFormControlModel,
     DynamicFormControlEvent,
-    DynamicFormRelationService
+    DynamicFormRelationService,
+    DynamicTemplateDirective
 } from "@ng2-dynamic-forms/core";
 
 export const DYNAMIC_FORM_UI_MATERIAL = "MATERIAL";
@@ -24,12 +25,13 @@ export class DynamicFormMaterialComponent extends DynamicFormControlComponent {
     @Input() bindId: boolean = true;
     @Input() controlGroup: FormGroup;
     @Input() model: DynamicFormControlModel;
-    @Input() nestedTemplate: TemplateRef<any>;
+    @Input() nestedTemplates: QueryList<any>;
 
     @Input()
     get showCharacterHint(): boolean {
         return !!(this._showCharacterCount && this.model["maxLength"] && this.characterCount);
     }
+
     set showCharacterHint(value: boolean) {
         this._showCharacterCount = value;
     }
@@ -38,7 +40,7 @@ export class DynamicFormMaterialComponent extends DynamicFormControlComponent {
     @Output() change: EventEmitter<DynamicFormControlEvent> = new EventEmitter<DynamicFormControlEvent>();
     @Output() focus: EventEmitter<DynamicFormControlEvent> = new EventEmitter<DynamicFormControlEvent>();
 
-    @ContentChild(TemplateRef) customTemplate;
+    @ContentChildren(DynamicTemplateDirective) templates: QueryList<any>;
 
     @ViewChild(MdCheckbox) mdCheckbox: MdCheckbox;
     @ViewChild(MdInputContainer) mdInputContainer: MdInputContainer;